bkp_17th_Oct_2025/mcp_server/pyproject.toml
[project]
name = "test-mcp"
version = "0.1.0"
description = "MCP Server With Azure AI Foudry"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-projects==1.0.0b11",
    "azure-core==1.34.0",
    "azure-identity==1.22.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "fastmcp>=0.9.0",
    "python-dotenv>=1.0.0"
]